move stack saving function to stack.py
